Update index.md CascadingProperty to CascadingParameter
Apologies if this is a misunderstanding, I'm new to Blazor (hence following the tutorial) but I think the CascadingProperty should actually be a CascadingParameter in this file.
Furthermore, I'm uncertain about the fifth point in the list `The property is public`. I have managed to cascade values onto a private property without naming the CascadingValue.